Split anime and manga examples in different files
